refactor: KEEP-356 group For Each node files under lib/workflow/nodes/for-each
Pure file relocation. The framework-detection regex (Workflow DevKit
"use workflow"/"use step" directives) is content-based, not path-based,
so this move is safe.

Moves:
- lib/for-each-concurrency.ts -> lib/workflow/nodes/for-each/concurrency.ts
- lib/for-each-utils.ts -> lib/workflow/nodes/for-each/utils.ts
- lib/iteration-grouping.ts -> lib/workflow/nodes/for-each/iteration-grouping.ts
- lib/steps/for-each.ts -> lib/workflow/nodes/for-each/step.ts
- lib/steps/__tests__/for-each.test.ts -> lib/workflow/nodes/for-each/__tests__/step.test.ts

8 import sites updated.
