+++ components/workflow/config/action-config.tsx
@@ -49,7 +49,7 @@ import {
   extractObjectPaths,
   resolveArraySourceElement,
   traverseDotPath,
-} from "@/lib/for-each-utils";
+} from "@/lib/workflow/nodes/for-each/utils";
 import {
   executionLogsAtom,
   lastExecutionLogsAtom,
